requirements.txt
openai==1.61.0
python-dotenv>=1.0.1
streamlit>=1.31.0
streamlit-ace>=0.1.1
pandas>=2.0.0
psutil>=5.9.0
browser-use
langchain
langchain-google-genai
